+++ pkg/domain/decoder_token.go
@@ -1,12 +1,10 @@
-package authentication
+package domain
 
 import (
 	"encoding/base64"
 	"strings"
 
 	"github.com/google/uuid"
-
-	"github.com/CMedrado/DesafioStone/pkg/domain"
 )
 
 // DecoderToken returns the ID that was inside an encrypted code
@@ -16,12 +14,12 @@ func DecoderToken(token string) (uuid.UUID, uuid.UUID, error) {
 
 	idInt, err := uuid.Parse(idString[3])
 	if err != nil {
-		return uuid.UUID{}, uuid.UUID{}, domain.ErrInvalidToken
+		return uuid.UUID{}, uuid.UUID{}, ErrInvalidToken
 	}
 
 	idToken, err := uuid.Parse(idString[4])
 	if err != nil {
-		return uuid.UUID{}, uuid.UUID{}, domain.ErrInvalidToken
+		return uuid.UUID{}, uuid.UUID{}, ErrInvalidToken
 	}
 
 	return idInt, idToken, nil
