+++ pkg/domain/usecases.go
@@ -18,13 +18,13 @@ type AccountUseCase interface {
 }
 
 type LoginUseCase interface {
-	AuthenticatedLogin(secret string, id entities.Account) (error, string)
+	AuthenticatedLogin(secret string, cpf string) (error, string)
 	GetTokenID(id uuid.UUID) (entities.Token, error)
 }
 
 type TransferUseCase interface {
-	GetTransfers(accountOrigin entities.Account, accountToken entities.Token, token string) ([]entities.Transfer, error)
-	CreateTransfers(ctx context.Context, accountOriginID uuid.UUID, accountToken entities.Token, token string, accountOrigin entities.Account, accountDestination entities.Account, amount int, accountDestinationIdUUID uuid.UUID) (error, uuid.UUID, entities.Account, entities.Account)
+	GetTransfers(token string) ([]entities.Transfer, error)
+	CreateTransfers(ctx context.Context, token string, amount int, accountDestinationIdUUID string) (error, uuid.UUID)
 	GetCountTransfer(ctx context.Context) (int64, error)
 	GetRankTransfer(ctx context.Context) ([]string, error)
 }
