+++ pkg/gateways/db/postgre/entries/change_balance.go
@@ -1,4 +1,4 @@
-package accounts
+package entries
 
 import (
 	"context"
