+++ pkg/gateways/db/postgre/entries/return_account_cpf.go
@@ -1,4 +1,4 @@
-package accounts
+package entries
 
 import (
 	"context"
