+++ pkg/gateways/db/postgre/entries/return_accounts.go
@@ -1,4 +1,4 @@
-package accounts
+package entries
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"github.com/CMedrado/DesafioStone/pkg/domain/entities"
 )
 
-func (a *Storage) ReturnAccounts() ([]entities.Account, error) {
+func (a Storage) ReturnAccounts() ([]entities.Account, error) {
 	statement := `SELECT * FROM accounts`
 
 	rows, err := a.pool.Query(context.Background(), statement)
