+++ pkg/gateways/db/postgre/entries/return_token_id.go
@@ -1,4 +1,4 @@
-package token
+package entries
 
 import (
 	"context"
