+++ pkg/gateways/db/postgre/entries/return_transfers.go
@@ -1,4 +1,4 @@
-package transfer
+package entries
 
 import (
 	"context"
