+++ pkg/gateways/db/postgre/entries/save_account.go
@@ -1,4 +1,4 @@
-package accounts
+package entries
 
 import (
 	"context"
