+++ pkg/gateways/db/postgre/entries/save_token.go
@@ -1,4 +1,4 @@
-package token
+package entries
 
 import (
 	"context"
