+++ pkg/gateways/db/postgre/entries/save_transfer.go
@@ -1,4 +1,4 @@
-package transfer
+package entries
 
 import (
 	"context"
