+++ pkg/gateways/db/postgre/entries/store.go
@@ -1,4 +1,4 @@
-package accounts
+package entries
 
 import (
 	"time"
@@ -6,6 +6,20 @@ import (
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v4/pgxpool"
 )
+
+type Token struct {
+	ID        uuid.UUID `json:"id"`
+	IdAccount uuid.UUID `json:"id_account"`
+	CreatedAt time.Time `json:"created_at"`
+}
+
+type Transfer struct {
+	ID                   uuid.UUID `json:"id"`
+	OriginAccountID      uuid.UUID `json:"origin_account_id"`
+	DestinationAccountID uuid.UUID `json:"destination_account_id"`
+	Amount               int       `json:"amount"`
+	CreatedAt            time.Time `json:"created_at"`
+}
 
 type Account struct {
 	ID        uuid.UUID `json:"id"`
