--- pkg/gateways/db/postgre/transfer/store.go
@@ -1,24 +1,0 @@
-package transfer
-
-import (
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v4/pgxpool"
-)
-
-type Transfer struct {
-	ID                   uuid.UUID `json:"id"`
-	OriginAccountID      uuid.UUID `json:"origin_account_id"`
-	DestinationAccountID uuid.UUID `json:"destination_account_id"`
-	Amount               int       `json:"amount"`
-	CreatedAt            time.Time `json:"created_at"`
-}
-
-type Storage struct {
-	pool *pgxpool.Pool
-}
-
-func NewStored(pool *pgxpool.Pool) *Storage {
-	return &Storage{pool: pool}
-}
