+++ pkg/gateways/http/authentication/handler.go
@@ -5,10 +5,9 @@ import (
 )
 
 type Handler struct {
-	account domain.AccountUseCase
-	login   domain.LoginUseCase
+	login domain.LoginUseCase
 }
 
-func NewHandler(accountUseCase domain.AccountUseCase, useCase domain.LoginUseCase) *Handler {
-	return &Handler{account: accountUseCase, login: useCase}
+func NewHandler(useCase domain.LoginUseCase) *Handler {
+	return &Handler{login: useCase}
 }
