+++ pkg/gateways/http/transfer/create_transfer_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/google/uuid"
 
-	domain2 "github.com/CMedrado/DesafioStone/pkg/domain"
+	"github.com/CMedrado/DesafioStone/pkg/domain"
 	"github.com/CMedrado/DesafioStone/pkg/domain/entities"
 	"github.com/CMedrado/DesafioStone/pkg/domain/transfer"
 )
@@ -106,8 +106,6 @@ func TestHandler_CreateTransfer(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			s := new(Handler)
 			s.transfer = &TransferUsecaseMock{}
-			s.account = &AccountUsecaseMock{}
-			s.login = &TokenUseCaseMock{}
 			bodyBytes := []byte(tc.body)
 			request, _ := http.NewRequest(tc.method, tc.path, bytes.NewReader(bodyBytes))
 			respondeRecorder := httptest.NewRecorder()
@@ -129,7 +127,7 @@ func TestHandler_CreateTransfer(t *testing.T) {
 type TransferUsecaseMock struct {
 }
 
-func (uc *TransferUsecaseMock) GetTransfers(_ entities.Account, _ entities.Token, token string) ([]entities.Transfer, error) {
+func (uc *TransferUsecaseMock) GetTransfers(token string) ([]entities.Transfer, error) {
 	time1, _ := time.Parse("2006-01-02T15:04:05.999999999Z07:00", "2021-07-20T15:17:25.933365Z")
 	if secret == token {
 		return []entities.Transfer{
@@ -143,26 +141,30 @@ func (uc *TransferUsecaseMock) GetTransfers(_ entities.Account, _ entities.Token
 		}, nil
 	}
 
+	if token == "MDIvMDgvMjAyMSAwOToyNzo0NDo2YjE5NDFkYi1jZTE3LTRmZmUtYTdlZC0yMeQ5M2E5MjZiYmM6YmQxODIxZTQtM2I5YS00M2RjLWJkZGUtNjBiM2QyMTRhYzdm" { // #nosec
+		return []entities.Transfer{}, domain.ErrInvalidToken
+	}
+
 	return []entities.Transfer{}, errors.New("given token is invalid")
 }
 
-func (uc TransferUsecaseMock) CreateTransfers(_ context.Context, accountOriginID uuid.UUID, _ entities.Token, token string, accountOrigin entities.Account, accountDestination entities.Account, amount int, accountDestinationIdUUID uuid.UUID) (error, uuid.UUID, entities.Account, entities.Account) {
+func (uc TransferUsecaseMock) CreateTransfers(_ context.Context, token string, amount int, accountDestinationIdUUID string) (error, uuid.UUID) {
 	if amount <= 0 {
-		return errors.New("given amount is invalid"), uuid.UUID{}, entities.Account{}, entities.Account{}
+		return errors.New("given amount is invalid"), uuid.UUID{}
 	}
 	if secret != token {
-		return errors.New("given token is invalid"), uuid.UUID{}, entities.Account{}, entities.Account{}
+		return domain.ErrInvalidToken, uuid.UUID{}
 	}
-	if accountOriginID == accountDestinationIdUUID {
-		return errors.New("given account is the same as the account destination"), uuid.UUID{}, entities.Account{}, entities.Account{}
+	if "6b1941db-ce17-4ffe-a7ed-22493a926bbc" == accountDestinationIdUUID {
+		return errors.New("given account is the same as the account destination"), uuid.UUID{}
 	}
-	if accountOrigin.Balance < amount {
-		return errors.New("given account without balance"), uuid.UUID{}, entities.Account{}, entities.Account{}
+	if 6000 < amount {
+		return errors.New("given account without balance"), uuid.UUID{}
 	}
-	if (accountDestination == entities.Account{}) {
-		return transfer.ErrInvalidDestinationID, uuid.UUID{}, entities.Account{}, entities.Account{}
+	if "75432539-c5ba-46d3-9690-44985b516da5" == accountDestinationIdUUID {
+		return transfer.ErrInvalidDestinationID, uuid.UUID{}
 	}
-	return nil, uuid.MustParse("c5424440-4737-4e03-86d2-3adac90ddd20"), accountOrigin, accountDestination
+	return nil, uuid.MustParse("c5424440-4737-4e03-86d2-3adac90ddd20")
 }
 
 func (uc *TransferUsecaseMock) GetCountTransfer(_ context.Context) (int64, error) {
@@ -171,122 +173,4 @@ func (uc *TransferUsecaseMock) GetCountTransfer(_ context.Context) (int64, error
 
 func (uc *TransferUsecaseMock) GetRankTransfer(_ context.Context) ([]string, error) {
 	return []string{}, nil
-}
-
-type TokenUseCaseMock struct {
-	AccountList AccountUsecaseMock
-}
-
-func (uc TokenUseCaseMock) AuthenticatedLogin(secret string, account entities.Account) (error, string) {
-	secretHash := domain2.CreateHash(secret)
-	if account == (entities.Account{}) {
-		return errors.New("given secret or CPF are incorrect"), ""
-	}
-	//if account.CPF != account.CPF {
-	//	return errors.New("given secret or CPF are incorrect"), ""
-	//}
-	if account.Secret != secretHash {
-		return errors.New("given secret or CPF are incorrect"), ""
-	}
-	return nil, "passou"
-}
-
-func (uc TokenUseCaseMock) GetTokenID(id uuid.UUID) (entities.Token, error) {
-	time1, _ := time.Parse("2006-01-02T15:04:05.999999999Z07:00", "2021-08-02T09:27:44.933365Z")
-	if id == uuid.MustParse("39a70a94-a82d-4db8-87ae-bd900c6a7c04") {
-		return entities.Token{
-			ID:        uuid.MustParse("39a70a94-a82d-4db8-87ae-bd900c6a7c04"),
-			IdAccount: uuid.MustParse("6b1941db-ce17-4ffe-a7ed-22493a926bbc"),
-			CreatedAt: time1,
-		}, nil
-	}
-	return entities.Token{}, nil
-}
-
-type AccountUsecaseMock struct {
-}
-
-func (uc AccountUsecaseMock) CreateAccount(_ context.Context, name string, cpf string, _ string, balance int) (uuid.UUID, error) {
-	if len(cpf) != 11 && len(cpf) != 14 {
-		return uuid.UUID{}, errors.New("given cpf is invalid")
-	}
-	if len(cpf) == 14 {
-		if string([]rune(cpf)[3]) == "." && string([]rune(cpf)[7]) == "." && string([]rune(cpf)[11]) == "-" {
-			if balance <= 0 {
-				return uuid.UUID{}, errors.New("given the balance amount is invalid")
-			}
-			if name == "Rafael" {
-				return uuid.MustParse("f7ee7351-4c96-40ca-8cd8-37434810ddfa"), nil
-			}
-			if name == "Lucas" {
-				return uuid.MustParse("a505b1f9-ac4c-45aa-be43-8614a227a9d4"), nil
-			}
-		} else {
-			return uuid.UUID{}, errors.New("given cpf is invalid")
-		}
-	}
-	if balance <= 0 {
-		return uuid.UUID{}, errors.New("given the balance amount is invalid")
-	}
-	if name == "Rafael" {
-		return uuid.MustParse("f7ee7351-4c96-40ca-8cd8-37434810ddfa"), nil
-	}
-	if name == "Lucas" {
-		return uuid.MustParse("a505b1f9-ac4c-45aa-be43-8614a227a9d4"), nil
-	}
-	return uuid.UUID{}, nil
-}
-
-func (uc AccountUsecaseMock) GetBalance(_ string) (int, error) {
-	return 0, nil
-}
-
-func (uc AccountUsecaseMock) GetAccounts() ([]entities.Account, error) {
-	time1, _ := time.Parse("2006-01-02T15:04:05.999999999Z07:00", "2021-08-02T09:41:46.813816-03:00")
-	return []entities.Account{
-		{
-			ID:        uuid.MustParse("a61227cf-a857-4bc6-8fcd-ad97cdad382a"),
-			Name:      "Rafael",
-			CPF:       "38453162093",
-			Secret:    "53b9e9679a8ea25880376080b76f98ad",
-			Balance:   6000,
-			CreatedAt: time1,
-		},
-		{
-			ID:        uuid.MustParse("6b1941db-ce17-4ffe-a7ed-22493a926bbc"),
-			Name:      "Lucas",
-			CPF:       "08131391043",
-			Secret:    "c74af74c69d81831a5703aefe9cb4199",
-			Balance:   5000,
-			CreatedAt: time1,
-		},
-	}, nil
-}
-
-func (uc AccountUsecaseMock) GetAccountID(id uuid.UUID) (entities.Account, error) {
-	account := entities.Account{}
-	accounts, _ := uc.GetAccounts()
-	for _, a := range accounts {
-		if a.ID == id {
-			account = a
-		}
-	}
-
-	return account, nil
-}
-
-func (uc *AccountUsecaseMock) UpdateBalance(_ entities.Account, _ entities.Account) error {
-	return nil
-}
-
-func (uc AccountUsecaseMock) GetAccountCPF(_ context.Context, cpf string) (entities.Account, error) {
-	account := entities.Account{}
-	accounts, _ := uc.GetAccounts()
-	for _, a := range accounts {
-		if a.CPF == cpf {
-			account = a
-		}
-	}
-
-	return account, nil
 }
