+++ pkg/gateways/http/transfer/handler.go
@@ -5,11 +5,9 @@ import (
 )
 
 type Handler struct {
-	account  domain.AccountUseCase
-	login    domain.LoginUseCase
 	transfer domain.TransferUseCase
 }
 
-func NewHandler(accountUseCase domain.AccountUseCase, loginUseCase domain.LoginUseCase, useCase domain.TransferUseCase) *Handler {
-	return &Handler{account: accountUseCase, login: loginUseCase, transfer: useCase}
+func NewHandler(useCase domain.TransferUseCase) *Handler {
+	return &Handler{transfer: useCase}
 }
