+++ pkg/gateways/http/transfer/list_transfer_test.go
@@ -46,8 +46,6 @@ func TestHandler_ListTransfers(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			s := new(Handler)
 			s.transfer = &TransferUsecaseMock{}
-			s.account = &AccountUsecaseMock{}
-			s.login = &TokenUseCaseMock{}
 			bodyBytes := []byte(tc.body)
 			request, _ := http.NewRequest(tc.method, tc.path, bytes.NewReader(bodyBytes))
 			respondRecorder := httptest.NewRecorder()
